Move zcash logic into separate module
